+++ checkall.py
@@ -20,7 +20,6 @@ def check(package: str, feature_list: list[str]):
     features = [["--all-features"], ["--no-default-features"]]
     features += [["--no-default-features", "--features", ",".join(features)] for features in allcombos]
 
-    abort = False
     for n, featureset in enumerate(features):
         for command in ["check", "clippy", "test"]:
             print(f">> [{package}] [{n:04d}/{len(features):04d}] cargo {command}", " ".join(featureset))
@@ -28,24 +27,13 @@ def check(package: str, feature_list: list[str]):
             # cargo test --all-targets skips doctests
             targets = [] if command == "test" else ["--all-targets"]
             ret = sp.run(["cargo", command, "--package", package] + targets + featureset)
-
-            try:
-                ret.check_returncode()
-
-            except sp.CalledProcessError:
-                abort = True
-
-            if abort:
-                break
-
-        if abort:
-            break
+            ret.check_returncode()
 
 
 def main():
     os.environ["QUICKCHECK_TESTS"] = "100000"
 
-    check("dxr", ["derive", "multicall", "i8", "nil"])
+    check("dxr", ["derive", "multicall", "i8", "nil", "chrono", "jiff", "time"])
     check("dxr_derive", [])
     check("dxr_client", ["default", "multicall", "reqwest", "default-tls", "native-tls", "rustls-tls"])
     check("dxr_server", ["default", "multicall", "axum"])
@@ -53,4 +41,8 @@ def main():
 
 
 if __name__ == "__main__":
-    main()
+    try:
+        main()
+    except Exception:
+        exit(1)
+