+++ dxr/src/values.rs
@@ -1,8 +1,6 @@
-/// Date & time format used by the XML-RPC `dateTime.iso8601` value type
-///
-/// This string represents the format of the timezone-unaware date & time format used by the XML-RPC
-/// `dateTime.iso8601` value type, to be used with [`chrono::NaiveDateTime::parse_from_str`].
-pub const XML_RPC_DATE_FORMAT: &str = "%Y%m%dT%H:%M:%S";
+mod datetime;
+pub use datetime::DateTime;
+pub use datetime::DateTimeParseError;
 
 mod ser_de;